[refactor] 파일 네이밍, 디렉토리 구조 리팩토링
