+++ notification/src/main/java/com/hubEleven/notification/slack/infrastructure/dto/response/SlackWebhookResponse.java
@@ -1,4 +1,4 @@
-package com.hubEleven.notification.slack.application.dto;
+package com.hubEleven.notification.slack.infrastructure.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
